+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   synchronization attempts.
 * Support gstreamer's udpsink as a data source for visualization (Mopidy).
 * Deprecate `visualizer_fifo_path` in favor of `visualizer_data_source`.
+* Don't run volume changing actions if there is no mixer.
 
 # ncmpcpp-0.8.2 (2018-04-11)
 * Help screen: fixed display of EoF keycode